Use .pem file if on windows
Instruct curl to use the needed .pem file if we are running on windows.
The code checks for the .pem file in the same location as the BlockIo
class file and if it's not found an exception will be thrown with
ionformation about downloading the .pem file and where to palce it.

This gets rid of the need to alter php configuration files in order to
get the library working on windows.